refactor: draft route 엔드포인트를 V2 컨트롤러에서 V1으로 이전 및 V2 컨트롤러 제거
# Conflicts:
#	backend/src/main/java/coursepick/coursepick/presentation/CourseV1WebController.java
#	backend/src/main/java/coursepick/coursepick/presentation/api/CourseWebApi.java
