+++ backend/src/main/java/coursepick/coursepick/presentation/CourseV1WebController.java
@@ -4,11 +4,9 @@ import coursepick.coursepick.application.CourseApplicationService;
 import coursepick.coursepick.application.dto.CoursesResponse;
 import coursepick.coursepick.domain.course.Coordinate;
 import coursepick.coursepick.domain.course.CourseFindCondition;
+import coursepick.coursepick.domain.course.DraftSegment;
 import coursepick.coursepick.presentation.api.CourseWebApi;
-import coursepick.coursepick.presentation.dto.CoordinateWebResponse;
-import coursepick.coursepick.presentation.dto.CourseWebResponse;
-import coursepick.coursepick.presentation.dto.CoursesWebResponse;
-import coursepick.coursepick.presentation.dto.CourseCreateWebRequest;
+import coursepick.coursepick.presentation.dto.*;
 import coursepick.coursepick.security.Login;
 import coursepick.coursepick.security.UserId;
 import jakarta.validation.Valid;
@@ -77,5 +75,12 @@ public class CourseV1WebController implements CourseWebApi {
     public String addCustomCourses(@Valid @RequestBody CourseCreateWebRequest request, @UserId String userId) {
         courseApplicationService.addCustomCourse(request.name(), request.toCoordinates(), userId);
         return "코스 추가 성공";
+    }
+
+    @Override
+    @PostMapping("/courses/draft/route")
+    public DraftRouteWebResponse findDraftRoute(@Valid @RequestBody FindDraftRouteWebRequest request) {
+        DraftSegment route = courseApplicationService.findDraftRoute(request.toCoordinates());
+        return DraftRouteWebResponse.of(route.coordinates(), route.length());
     }
 }
